work on the lab

+++ include/goya/primitives.hpp
@@ -7,7 +7,7 @@
 
 namespace goya {
 
-using IndexType = std::size_t;
+using IndexType = std::uint32_t;
 
 using Vertex3d = glm::vec3;
 using Face = std::vector<IndexType>;
